refactor(cli): extract search.py action dispatch + approval flow
B14 of the Round 2 plan. Moves the three action-flow functions out
of amx/cli_support/commands/search.py into a focused
amx/cli_support/_search_actions.py:

- _run_search_action (dispatch sync_catalog / refresh_code_evidence /
  analyze_table from a search result)
- _run_approved_search_actions (interactive picker over actions the
  user is allowed to approve)
- _sync_cached_code_evidence (push cached code-scan evidence into
  the search catalog)

_run_search_action cross-calls helpers that still live in search.py
(_answer_scope, _interactive_sync_scope, _sync_db_scope); those calls
go via lazy import inside the function so tests monkeypatching
the parent module keep working.

search.py: 1283 -> 1117 LOC (-166, -13%).
